// src/core/schemas/enums.ts
export enum MotionEffect {
    ZOOM_IN_TOP_LEFT = "ZOOM_IN_TOP_LEFT",
    ZOOM_IN_TOP_RIGHT = "ZOOM_IN_TOP_RIGHT",
    ZOOM_IN_BOTTOM_LEFT = "ZOOM_IN_BOTTOM_LEFT",
    ZOOM_IN_BOTTOM_RIGHT = "ZOOM_IN_BOTTOM_RIGHT",
    ZOOM_IN_CENTER = "ZOOM_IN_CENTER",
    ZOOM_OUT_TOP_LEFT = "ZOOM_OUT_TOP_LEFT",
    ZOOM_OUT_TOP_RIGHT = "ZOOM_OUT_TOP_RIGHT",
    ZOOM_OUT_BOTTOM_LEFT = "ZOOM_OUT_BOTTOM_LEFT",
    ZOOM_OUT_BOTTOM_RIGHT = "ZOOM_OUT_BOTTOM_RIGHT",
    ZOOM_OUT_CENTER = "ZOOM_OUT_CENTER",
    // SLIDE_LEFT = "SLIDE_LEFT",
    // SLIDE_RIGHT = "SLIDE_RIGHT",
    // SLIDE_UP = "SLIDE_UP",
    // SLIDE_DOWN = "SLIDE_DOWN",
}

export enum EffectEffect {
    BLUR = "BLUR",
    GRAYSCALE = "GRAYSCALE",
}
export enum ConnectEffect {
    SPLIT = "SPLIT",
    CONNECT = "CONECT",
}
export type EffectType = MotionEffect | EffectEffect | ConnectEffect;

export enum OutputAspectRatio {
    RATIO_16_9 = "16:9",
    RATIO_9_16 = "9:16",
    RATIO_1_1 = "1:1",
    RATIO_4_5 = "4:5",
    RATIO_4_3 = "4:3",
}
export enum OutputFormat {
    MP4 = "mp4",
    GIF = "gif",
    JPG = "jpg",
    PNG = "png",
    BMP = "bmp",
    MP3 = "mp3",
}
export enum OutputFPS {
    FPS_12 = 12,
    FPS_15 = 15,
    FPS_24 = 24,
    FPS_23_976 = 23.976,
    FPS_25 = 25,
    FPS_29_97 = 29.97,
    FPS_30 = 30,
    FPS_60 = 60,
}
// ...
export enum XFadeTransition {
    FADE = "fade",
    FADE_BLACK = "fadeblack",
    FADE_WHITE = "fadewhite",
    DISTANCE = "distance",
    WIPE_LEFT = "wipeleft",
    WIPE_RIGHT = "wiperight",
    WIPE_UP = "wipeup",
    WIPE_DOWN = "wipedown",
    SLIDE_LEFT = "slideleft",
    SLIDE_RIGHT = "slideright",
    SLIDE_UP = "slideup",
    SLIDE_DOWN = "slidedown",
    SMOOTH_LEFT = "smoothleft",
    SMOOTH_RIGHT = "smoothright",
    SMOOTH_UP = "smoothup",
    SMOOTH_DOWN = "smoothdown",
    RECT_CROP = "rectcrop",
    CIRCLE_CROP = "circlecrop",
    CIRCLE_CLOSE = "circleclose",
    CIRCLE_OPEN = "circleopen",
    HORZ_CLOSE = "horzclose",
    HORZ_OPEN = "horzopen",
    VERT_CLOSE = "vertclose",
    VERT_OPEN = "vertopen",
    DIAG_BL = "diagbl",
    DIAG_BR = "diagbr",
    DIAG_TL = "diagtl",
    DIAG_TR = "diagtr",
    HL_SLICE = "hlslice",
    HR_SLICE = "hrslice",
    VU_SLICE = "vuslice",
    VD_SLICE = "vdslice",
    DISSOLVE = "dissolve",
    PIXELIZE = "pixelize",
    RADIAL = "radial",
    H_BLUR = "hblur",
    WIPE_TL = "wipetl",
    WIPE_TR = "wipetr",
    WIPE_BL = "wipebl",
    WIPE_BR = "wipebr",
    FADEGRAYS = "fadegrays",
    SQUEEZE_V = "squeezev",
    SQUEEZE_H = "squeezeh",
    ZOOM_IN = "zoomin",
}